+++ system/config/constants.php
@@ -13,7 +13,7 @@
  * Core version
  */
 define('VERSION', '3.5');
-define('BUILD', '17');
+define('BUILD', '18');
 define('LONG_TERM_SUPPORT', true);
 
 
@@ -23,7 +23,7 @@ define('LONG_TERM_SUPPORT', true);
  * Version numbers are added to style sheets and JavaScript files to make
  * the web browser reload those resources after a Contao update.
  */
-define('ACE', '1.1.8');
+define('ACE', '1.1.9');
 define('CSS3PIE', '1.0.0');
 define('DROPZONE', '3.12.0');
 define('HIGHLIGHTER', '3.0.83');